requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.3.0
django-rest-auth==0.9.5
django-six==1.0.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
Pillow==7.1.1